Fix: Refactor: Extract GpuAnalyzer struct to src/analysis/gpu/analyzer.rs (#273)
Extract GpuAnalyzer struct and GpuEvaluator trait from the monolithic
implementation.rs file to a dedicated analyzer.rs module under src/analysis/gpu/.

This change moves approximately 1800 lines of GPU-related code:
- GpuAnalyzer struct definition and all its methods
- GpuEvaluator trait definition with both implementations
- GPU compute pipeline builders (helpful, harmful, relu, activation, bias)
- Batch evaluation methods and result merging
- Shader constants (GPU_MAX_BATCH_ALLOC_BYTES)
- Buffer initialization and polling utilities

The refactoring maintains backwards compatibility through re-exports in
mod.rs and adds a has_gpu() method to replace direct device field access.

Part of issue #185 (implementation.rs monolith refactoring).
